scribe: initial VSD Performance read API and monitoring
Performance counters and averages have been added to the VSDEngine.
These stats are available via the new thrift call VSDPerformanceRead().

Initial data collected is 1,5,15 min decaying averages of IOPS,MBPS,
IOSize,Latency.  Also total data read/written is collected.

Reviewed-on: http://gerrit.lab.local:8080/3390
Tested-by: build
Smoketests: build

+++ ExponentiallyDampedMovingAverage.h
@@ -44,12 +44,12 @@ namespace Forte
                 Forte::AutoUnlockMutex lock(mLock);
                 int64_t current = mValueGetter() << mPrecision;
                 mValue *= mExp;
-                mValue += current * ((1 << mPrecision) - mExp);
+                mValue += current * ((1ULL << mPrecision) - mExp);
                 mValue >>= mPrecision;
             }
             Forte::Mutex mLock;
             int64_t mValue;
-            unsigned mPrecision;
+            uint64_t mPrecision;
             int64_t mExp;
             Forte::Timespec mUpdateInterval;
             boost::function<int64_t(void)> mValueGetter;
